Revert "More nextjs style"
This reverts commit 780b9884bbeb3a6879b7862096b88219220ab403.
