+++ src/components/changelog/months.tsx
@@ -1,5 +1,5 @@
 'use client';
-
+import {Fragment} from 'react';
 import {ReadonlyURLSearchParams, usePathname, useSearchParams} from 'next/navigation';
 
 export default function Tags({months}: {months: any}) {
@@ -21,17 +21,20 @@ export default function Tags({months}: {months: any}) {
   };
 
   return (
-    <ul>
-      {months.map((month, index) => (
-        <li key={index}>
-          <a
-            className={`text-primary hover:text-purple-900 hover:font-extrabold ${before === month ? 'underline' : ''}`}
-            href={createPageURL(month)}
-          >
-            {month}
-          </a>
-        </li>
-      ))}
-    </ul>
+    <Fragment>
+      <h3 className="text-1xl text-primary font-semibold mb-2">Jump to:</h3>
+      <ul>
+        {months.map((month, index) => (
+          <li key={index}>
+            <a
+              className={`text-primary hover:text-purple-900 hover:font-extrabold ${before === month ? 'underline' : ''}`}
+              href={createPageURL(month)}
+            >
+              {month}
+            </a>
+          </li>
+        ))}
+      </ul>
+    </Fragment>
   );
 }
